requirements.txt
fastmcp

pydantic>=2.5.0
pydantic-settings>=2.1.0
anyio>=4.3.0

# --- Parsing ---
pymupdf>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.0
polars>=0.20.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
markdownify>=0.12.0
python-magic>=0.4.27
chardet>=5.2.0
markdown-it-py

# --- Images ---
Pillow>=10.3.0

# --- Cache ---
cachetools>=5.3.0
redis>=5.0.0
msgpack>=1.0.0

# --- Search ---
rank-bm25>=0.2.2

# --- Observability ---
structlog>=24.0.0
sentry-sdk>=1.40.0

# --- Dev / Test ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-benchmark>=4.0.0
pytest-cov>=4.0.0
coverage>=7.0.0

# 🔒 CRITICAL PINS
httpx==0.27.2
protobuf>=5.0,<7
posthog<4

# Optional but needed if using traceloop
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-exporter-otlp-proto-grpc>=1.28.0
opentelemetry-instrumentation>=0.55b0
opentelemetry-semantic-conventions>=0.55b0
opentelemetry-instrumentation-urllib3>=0.50b0

ruff>=0.4.0
mypy>=1.9.0